Misc doc-related changes
numbers.py: Precision; fix spelling, remove hard line-wrap and make
            render in LaTeX better
symbols.py: Need \n after </dl>. Hash was not doctest'd properly
combinatorial.py: Add Subset link; shorten Subset test output,
constants.py: better Infinity exmaples, link to DirectedInfinity
orderstats. WMA -> WMA link (when there are no other links)
expression_tests.py: move and *Q here.
systemsymbols: More symbols
