+++ mathics/builtin/testing_expressions/expression_tests.py
@@ -1,6 +1,7 @@
 """
 Expression Tests
 """
+from mathics.core.atoms import Integer0, Integer1, IntegerM1
 from mathics.core.builtin import Builtin, PatternError, Test
 from mathics.core.evaluation import Evaluation
 from mathics.core.symbols import SymbolFalse, SymbolTrue
@@ -66,4 +67,100 @@ class MatchQ(Builtin):
             return SymbolFalse
         except PatternError as e:
             evaluation.message(e.name, e.tag, *(e.args))
+            return SymbolFalse
+
+
+class Order(Builtin):
+    """
+    <url>:WMA link:https://reference.wolfram.com/language/ref/Order.html</url>
+
+    <dl>
+      <dt>'Order[$x$, $y$]'
+      <dd>returns a number indicating the canonical ordering of $x$ and $y$. \
+         1 indicates that $x$ is before $y$, and -1 that $y$ is before $x$. \
+         0 indicates that there is no specific ordering. Uses the same order \
+         as 'Sort'.
+    </dl>
+
+    >> Order[7, 11]
+     = 1
+
+    >> Order[100, 10]
+     = -1
+
+    >> Order[x, z]
+     = 1
+
+    >> Order[x, x]
+     = 0
+    """
+
+    summary_text = "order expressions"
+
+    def eval(self, x, y, evaluation: Evaluation):
+        "Order[x_, y_]"
+        if x < y:
+            return Integer1
+        elif x > y:
+            return IntegerM1
+        else:
+            return Integer0
+
+
+class OrderedQ(Builtin):
+    """
+    <url>
+    :WMA link:
+    https://reference.wolfram.com/language/ref/OrderedQ.html</url>
+
+    <dl>
+      <dt>'OrderedQ[{$a$, $b$}]'
+      <dd>is 'True' if $a$ sorts before $b$ according to canonical
+        ordering.
+    </dl>
+
+    >> OrderedQ[{a, b}]
+     = True
+    >> OrderedQ[{b, a}]
+     = False
+    """
+
+    summary_text = "test whether elements are canonically sorted"
+
+    def eval(self, expr, evaluation: Evaluation):
+        "OrderedQ[expr_]"
+
+        for index, value in enumerate(expr.elements[:-1]):
+            if expr.elements[index] <= expr.elements[index + 1]:
+                continue
+            else:
+                return SymbolFalse
+        return SymbolTrue
+
+
+# Note not in WMA anymore
+class PatternsOrderedQ(Builtin):
+    """
+    <dl>
+      <dt>'PatternsOrderedQ[$patt1$, $patt2$]'
+      <dd>returns 'True' if pattern $patt1$ would be applied before
+        $patt2$ according to canonical pattern ordering.
+    </dl>
+
+    >> PatternsOrderedQ[x__, x_]
+     = False
+    >> PatternsOrderedQ[x_, x__]
+     = True
+    >> PatternsOrderedQ[b, a]
+     = True
+    """
+
+    summary_text = "test whether patterns are canonically sorted"
+
+    def eval(self, p1, p2, evaluation: Evaluation):
+        "PatternsOrderedQ[p1_, p2_]"
+
+        if p1.get_sort_key(True) <= p2.get_sort_key(True):
+            return SymbolTrue
+        else:
             return SymbolFalse
