+++ mathics/core/systemsymbols.py
@@ -172,6 +172,7 @@ SymbolNumberQ = Symbol("System`NumberQ")
 SymbolNumericQ = Symbol("System`NumericQ")
 SymbolO = Symbol("System`O")
 SymbolOpacity = Symbol("System`Opacity")
+SymbolOperate = Symbol("System`Operate")
 SymbolOptionValue = Symbol("System`OptionValue")
 SymbolOptional = Symbol("System`Optional")
 SymbolOptions = Symbol("System`Options")
@@ -230,6 +231,7 @@ SymbolSin = Symbol("System`Sin")
 SymbolSinh = Symbol("System`Sinh")
 SymbolSlot = Symbol("System`Slot")
 SymbolSparseArray = Symbol("System`SparseArray")
+SymbolSortBy = Symbol("System`SortBy")
 SymbolSplit = Symbol("System`Split")
 SymbolSqrt = Symbol("System`Sqrt")
 SymbolSqrtBox = Symbol("System`SqrtBox")
