+++ mathics/doc/documentation/1-Manual.mdoc
@@ -1,8 +1,7 @@
 
 <chapter title="Introduction">
 
-\Mathics---to be pronounced like "Mathematics" without the "emat"---is
-a <url>:computer algebra
+\Mathics is a <url>:computer algebra
 system:https://en.wikipedia.org/wiki/Computer_algebra_system</url>. It
 is a free, open-source alternative to \Mathematica or the \Wolfram
 Language. However, \Mathics builds around and on top of the Python
